Refactor to create a more maintainable structure
