+++ cads_adaptors/adaptors/legacy.py
@@ -1,0 +1,18 @@
+
+from typing import BinaryIO
+
+from cads_adaptors.adaptors import Adaptor, Request
+
+class LegacyCdsAdaptor(Adaptor):
+    def retrieve(self, request: Request) -> BinaryIO:
+        import cdsapi
+
+        # parse input options
+        collection_id = self.config.pop("collection_id", None)
+        if not collection_id:
+            raise ValueError("collection_id is required in request")
+
+        # retrieve data
+        client = cdsapi.Client(self.config["url"], self.config["key"], retry_max=1)
+        result_path = client.retrieve(collection_id, request).download()
+        return open(result_path, "rb")
