+++ cads_adaptors/adaptors/url.py
@@ -1,0 +1,26 @@
+
+from typing import BinaryIO
+
+from cads_adaptors import mapping
+from cads_adaptors.adaptors import Adaptor, Request
+
+
+class UrlAdaptor(Adaptor):
+    def retrieve(self, request: Request) -> BinaryIO:
+        from cads_adaptors.tools import url_tools
+
+        data_format = request.pop("format", "zip")
+
+        if data_format not in {"zip", "tgz"}:
+            raise ValueError(f"{data_format=} is not supported")
+
+        mapped_request = mapping.apply_mapping(request, self.mapping)  # type: ignore
+
+        requests_urls = url_tools.requests_to_urls(
+            mapped_request, patterns=self.config["patterns"]
+        )
+
+        path = url_tools.download_from_urls(
+            [ru["url"] for ru in requests_urls], data_format=data_format
+        )
+        return open(path, "rb")
